requirements.txt
ollama~=0.6.1
pillow~=12.1.1